+++ LED.h
@@ -8,6 +8,6 @@
 #include <Ticker.h>
 
 /* Function prototypes */
-void LED_Blink(int TimeSeconds);
+void LED_Blink(float TimeSeconds);
 
 #endif /* LED_H */
